+++ CHANGELOG.md
@@ -1,9 +1,14 @@
+## 22.10.1
+NEW FEATURES:
+* resource/cvo_gcp HA: support add, rename and delete SVMs.
+
 ## 22.10.0
 NEW FEATURES:
 * resource/connector_snapmirror: support fsx as a source for snapmirror relationship with cvo.
 * resource/cvo_aws: add `retries` parameter to increase wait time when creating CVO.
 * resource/cvo_azure: add `retries` parameter to increase wait time when creating CVO.
 * resource/cvo_gcp: add `retries` parameter to increase wait time when creating CVO.
+* resoruce/cvs for AWS, AZURE and GCP: add `svm_name` an optional parameter. The modification is supported.
 
 NEW ENHANCEMENTS:
 * resource/connector_azure: display the deployed virtual machine principal_id in state file on the connector azure creation.